backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pymupdf
python-docx
ollama
python-dotenv
pydantic
httpx
pydantic-settings
